엔티티 일부 수정 및 검색 알고리즘 추가

+++ server/demo/src/main/java/com/example/demo/Diet/DietController.java
@@ -17,7 +17,7 @@ public class DietController {
 
     @PostMapping("/api/searchDietList")
     public ResponseEntity<List<DietDto>> getDietList(@RequestParam("BMR") double BMR,
-                                                     @RequestParam("activeCoef") double activeCoef){
+                                                     @RequestParam("activeCoef") int activeCoef){
         return ResponseEntity.ok(dietService.getDietList(BMR, activeCoef));
     }
